dev: rename ghostos_apps.py to ghostos_bootstrap.py
