+++ ghostos/prototypes/README.md
@@ -1,0 +1,3 @@
+# Prototypes
+
+The prototypes of `GhostOS` applications.
