--- ghostos/prototypes/streamlitapp/patches/streamlit_pydantic_patch.py
@@ -1,11 +1,0 @@
-from pydantic_settings import BaseSettings
-import pydantic
-
-pydantic.BaseSettings = BaseSettings
-
-# after pydantic patch
-import streamlit_pydantic
-
-__all__ = [
-    'streamlit_pydantic'
-]
